+++ packages/jaspr/lib/html.dart
@@ -1,4 +1,0 @@
-@Deprecated("Use 'package:jaspr/jaspr.dart' instead.")
-library html;
-
-export 'jaspr.dart' hide runApp;