+++ packages/jaspr/lib/jaspr.dart
@@ -3,21 +3,21 @@ export 'package:meta/meta.dart'
 
 export 'src/components/async.dart';
 export 'src/components/basic.dart';
+export 'src/components/html/html.dart';
 export 'src/components/style.dart';
 export 'src/foundation/annotations.dart';
 export 'src/foundation/basic_types.dart';
 export 'src/foundation/binding.dart';
 export 'src/foundation/change_notifier.dart';
 export 'src/foundation/constants.dart';
-export 'src/foundation/events.dart';
+export 'src/foundation/events/events.dart';
 export 'src/foundation/object.dart';
 export 'src/foundation/options.dart';
 export 'src/foundation/scheduler.dart';
+export 'src/foundation/styles/styles.dart';
 export 'src/foundation/sync.dart';
 export 'src/foundation/synchronous_future.dart';
 export 'src/framework/framework.dart';
 export 'src/stub/run_app_stub.dart'
     if (dart.library.html) 'src/browser/run_app.dart'
     if (dart.library.io) 'src/server/run_app.dart' show runApp;
-export 'src/ui/html/html.dart';
-export 'src/ui/styles.dart';