+++ packages/jaspr/lib/src/foundation/events/events.dart
@@ -1,8 +1,8 @@
-import '../ui/html/html.dart';
-import 'basic_types.dart';
-import 'events/events_web.dart' if (dart.library.ffi) 'events/events_vm.dart';
+import '../../components/html/html.dart';
+import '../basic_types.dart';
+import 'events_web.dart' if (dart.library.ffi) 'events/events_vm.dart';
 
-export 'events/events_web.dart' if (dart.library.ffi) 'events/events_vm.dart'
+export 'events_web.dart' if (dart.library.ffi) 'events/events_vm.dart'
     hide InputElement, TextAreaElement, SelectElement;
 
 typedef EventCallback = void Function(Event event);